Remove the depth from the key of RadixSort。

+++ cuda_rasterizer/rasterizer_impl.h
@@ -57,8 +57,8 @@ namespace CudaRasterizer
 	struct GS_BinningState
 	{
 		size_t sorting_size;
-		uint64_t* point_list_keys_unsorted;
-		uint64_t* point_list_keys;
+		uint32_t* point_list_keys_unsorted;
+		uint32_t* point_list_keys;
 		uint32_t* point_list_unsorted;
 		uint32_t* point_list;
 		char* list_sorting_space;
